add method assignSkierToPiste

+++ src/main/java/tn/esprit/feresski/entities/Skieur.java
@@ -1,5 +1,6 @@
 package tn.esprit.feresski.entities;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,11 +18,12 @@ public class Skieur {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY )
     int id;
-    Integer numSkieur;
+    Long numSkieur;
     String nomS;
     String prenomS;
     LocalDate dateNaissance;
     String ville;
+//@JsonIgnore
 @ManyToMany(mappedBy = "skieurs")
 Set <Piste>pistes;
